+++ src/db/migrations/migrations.js
@@ -10,6 +10,7 @@ import m0005 from './0005_list_look.sql';
 import m0006 from './0006_item_price.sql';
 import m0007 from './0007_shop_total.sql';
 import m0008 from './0008_wishes.sql';
+import m0009 from './0009_pantry.sql';
 
   export default {
     journal,
@@ -22,7 +23,8 @@ m0004,
 m0005,
 m0006,
 m0007,
-m0008
+m0008,
+m0009
     }
   }